refactor: Move exceptions to data.exception
