+++ src/winter/data/exception/InvalidPackageNameException.java
@@ -1,4 +1,4 @@
-package winter.exception;
+package winter.data.exception;
 
 public class InvalidPackageNameException extends Exception {
     public InvalidPackageNameException(String msg) {
