+++ src/winter/data/exception/MappingNotFoundException.java
@@ -1,4 +1,4 @@
-package winter.exception;
+package winter.data.exception;
 
 public class MappingNotFoundException extends Exception {
     public MappingNotFoundException() {
